Gabi/Docker/Task5-FullStack-Compose/backend/requirements.txt
Flask==3.0.3
psycopg2-binary==2.9.9
redis==5.0.7